feat(client): use Navigation API for routing when available
Implement the client router against the Navigation API and keep the
existing History API + click/popstate implementation as a fallback.
Feature detection gates on `NavigateEvent.prototype.sourceElement` —
its presence implies a sufficiently complete implementation.

On the Navigation API path, a single `navigate` listener replaces the
global click/popstate/hashchange listeners:

  - link-click filters (`.vp-raw`, button wrappers, `download`/`target`,
    cross-origin, non-HTML paths) are reproduced via `sourceElement`
  - `scroll: 'manual'` + `focusReset: 'manual'` preserve the existing
    `scrollTo` helper's a11y focus handling
  - scroll position is persisted on the outgoing entry via
    `updateCurrentEntry` and restored on `traverse`
  - fragment-directive (`#:~:text=…`) hash navigations fall through to
    native browser handling
  - `router.go` calls `navigation.navigate(...).finished`; a `__vpFromGo`
    info flag prevents `onBeforeRouteChange` from running twice
  - a `skipInternalNavigate` guard prevents re-entering the listener
    during the same-document `history.replaceState` URL fix-up in
    `loadPage`

Split the router into a small folder with two composable strategies
sharing the same `{ go, replaceUrl }` shape:

  src/client/app/router/
    router.ts         - public API, picks a strategy, wires loadPage
    shared.ts         - types, fakeHost, normalizeHref, scrollTo, …
    legacy.ts         - createLegacyRouterStrategy (History API)
    navigationApi.ts  - hasNavigationApi + Navigation API strategy

`loadPage` delegates same-document URL fix-up to `strategy.replaceUrl`
so each strategy handles its own re-entry semantics.
